Try covent position of btn exit dialog to position of btn previous btn and upload border of add image btn :100:

+++ Server/app/image.js
@@ -27,7 +27,7 @@ module.exports = app => {
             }
             
             imgurUploader(fs.readFileSync(filename), {title: `Image`}).then(data => {
-                fs.unlink(filename);
+                if(filename) fs.unlink(filename);
                 filename = '';
                 return res.send(data.link);
             })
